Deciding to remove Thunderstore Version checking of Plugins and only have a command for listing the current version of plugins after deciding potentially too many issues
